Add function to find an exact match on any key of an entry
A few things

This is a mess, because I don't know what I'm doing and have no idea how
to properly map/filter. Should refactor eventually but for now it's
completely functional and does work fine.

It'll loop over each `entry` block, and do an exact string match on each
key. The only exception is if the key is `tags`, in which case we need
to do an exact string match on each tag in the array.

Super naive but quick and dirty, until I have a better grasp on how to
do this idiomatically